Use gradle and reflection
