+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Fri Oct 26 09:12:20 EEST 2018
+gradle.version=4.10.2
