--- src/eu/insertcode/wordgames/compatibility/Compatibility.java
@@ -1,7 +1,0 @@
-package eu.insertcode.wordgames.compatibility;
-
-import org.bukkit.entity.Player;
-
-public interface Compatibility {
-	void sendJson(Player p, String json);
-}
