--- src/eu/insertcode/wordgames/compatibility/Compatibility_1_13_R2.java
@@ -1,17 +1,0 @@
-package eu.insertcode.wordgames.compatibility;
-
-import net.minecraft.server.v1_13_R2.IChatBaseComponent;
-import net.minecraft.server.v1_13_R2.PacketPlayOutChat;
-
-import org.bukkit.craftbukkit.v1_13_R2.entity.CraftPlayer;
-import org.bukkit.entity.Player;
-
-public class Compatibility_1_13_R2 implements Compatibility {
-	
-	@Override
-	public void sendJson(Player p, String json) {
-		IChatBaseComponent component = IChatBaseComponent.ChatSerializer.a(json);
-		PacketPlayOutChat packet = new PacketPlayOutChat(component);
-		((CraftPlayer) p).getHandle().playerConnection.sendPacket(packet);
-	}
-}
