--- src/eu/insertcode/wordgames/compatibility/Compatibility_1_8_R3.java
@@ -1,17 +1,0 @@
-package eu.insertcode.wordgames.compatibility;
-
-import net.minecraft.server.v1_8_R3.IChatBaseComponent;
-import net.minecraft.server.v1_8_R3.PacketPlayOutChat;
-
-import org.bukkit.craftbukkit.v1_8_R3.entity.CraftPlayer;
-import org.bukkit.entity.Player;
-
-public class Compatibility_1_8_R3 implements Compatibility {
-	
-	@Override
-	public void sendJson(Player p, String json) {
-		IChatBaseComponent component = IChatBaseComponent.ChatSerializer.a(json);
-		PacketPlayOutChat packet = new PacketPlayOutChat(component);
-		((CraftPlayer) p).getHandle().playerConnection.sendPacket(packet);
-	}
-}
