+++ src/main/java/eu/insertcode/wordgames/AutoStart.java
