+++ src/main/java/eu/insertcode/wordgames/CommandHandler.java
