+++ src/main/java/eu/insertcode/wordgames/ConfigManager.java
