+++ src/main/java/eu/insertcode/wordgames/Permission.java
