+++ src/main/java/eu/insertcode/wordgames/games/CalculateGame.java
