+++ src/main/java/eu/insertcode/wordgames/games/LongWordGame.java
