+++ src/main/java/eu/insertcode/wordgames/games/TimedGame.java
