+++ src/main/resources/config.yml
@@ -1,0 +1,71 @@
+#Need help with the configuration? There is a detailed explanation on https://www.spigotmc.org/resources/wordgames.33080/
+
+####################
+# General Settings #
+####################
+gameOptions:
+  # Whether the game should automatically be stopped after er certain amount of time. Does not apply to TimedGame.
+  autoStop:
+    enabled: false
+    timerInSeconds: 2750
+  #This is the command that will be executed once someone won.
+  rewardCommandSyntax: 'give {username} {reward} {amount}'
+  #rewardCommandSyntax: 'money give {username} {amount}' - You could also use this to give money instead.
+  #
+  #Amount of games that can play at the same time. Set to 0 to remove the limit.
+  maxPlayingGames: 1
+  scheduler:
+    #The wait between each game and/or each message.
+    timerInSeconds: 900
+  unmute:
+    #This is how many percent of the characters should be muted in the unmute game type.
+    percentageOfCharactersToMute: 25
+  timed:
+    #This is how many seconds the users have to type the word in the timed game type.
+    secondsToType: 10
+
+
+######################
+# autoStart settings #
+######################
+# If enabled, this will automatically start minigames.
+autoStart:
+  enabled: false
+  #
+  #Amount of players that should be online for the autostart to run.
+  minimumPlayers: 4
+  #
+  #All possible rewards the autostart will use. The syntax should be [amount] <reward>
+  rewards:
+  - 'diamond'
+  - 'ironingot'
+  - '4 ironingot'
+  - 'egg'
+  - 'itemframe'
+  - 'painting'
+  - 'beacon'
+  - 'spawn_egg:120'
+  #
+  #All the possible words the autostart will use.
+  words:
+  - 'WordGames+'
+  - 'insertCode'
+  - 'Minecraft'
+  - 'enderman'
+  - 'enderdragon'
+  - 'creeper'
+  - 'spider'
+  - 'bookshelf'
+  - 'redstone'
+  - 'elytra'
+  - 'pufferfish'
+  - 'enchanted'
+  - 'apocalypse'
+  - 'nether'
+  - 'homework'
+  - 'administrator'
+  - 'alpaca'
+  - 'message'
+  - 'player'
+  - 'another'
+  - 'word'
